sc2f/requirements.txt
brotli==1.1.0
ConfigParser==7.2.0
cryptography==46.0.2
Cython==3.1.4
dl==0.1.0
docutils==0.22.2
einops==0.8.1
HTMLParser==0.0.2
importlib_metadata==8.0.0
keyring==25.6.0
lockfile==0.12.2
lxml==6.0.2
matplotlib==3.10.6
numpy==2.3.3
opencv_python==4.12.0.88
ordereddict==1.1
Pillow==11.3.0
protobuf==6.32.1
pycocotools==2.0.10
pyOpenSSL==25.3.0
scipy==1.16.2
thop==0.1.1.post2209072238
thread==2.0.5
timm==1.0.20
toml==0.10.2
torch==2.8.0+cu126
torchvision==0.23.0+cu126
tornado==6.5.2
tqdm==4.67.1
xformers==0.0.32.post2
xlwt==1.3.0
xmlrpclib==1.0.1
zipp==3.19.2
